backend/requirements.txt
gradio==4.19.2
ultralytics==8.0.220
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0